requirements.txt
pyyaml>=6.0
skyfield>=1.48
jplephem>=2.21
numpy>=1.26
pytest>=8.2
